+++ src/heif/boxes/item_location.rs
@@ -375,7 +375,7 @@ ItemLocationBox
         data_start:  u64,
         data_length: u64
     )
-    -> (u32, usize)
+    -> (u32, u64)
     {
         // Determine largest iloc ID so far
         let old_largest_id = self.items
@@ -409,7 +409,7 @@ ItemLocationBox
         // TODO: make this more efficient by only computing how much memory is
         // needed, not by actually serializing (and thus, allocating memory)
         let old_box_size = self.header.get_box_size();
-        let new_box_size = self.serialize().len();
+        let new_box_size = self.serialize().len() as u64;
         self.header.set_box_size(new_box_size);
 
         return (
